replace int to long long

+++ 4. String/01.Z function.cpp
@@ -1,8 +1,8 @@
-vector<int> z_function(string& s) {
-    int n = s.size();
+vector<ll> z_function(string& s) {
+    ll n = s.size();
     vector<int> z(n);
-    int l = 0, r = 0;
-    for(int i = 1; i < n; i++) {
+    ll l = 0, r = 0;
+    for(ll i = 1; i < n; i++) {
         if(i < r) {
             z[i] = min(r - i, z[i - l]);
         }
